Problem 1: limit the output :checkered_flag:
There are couple of ways to manage concurrency. Channels are the most
common one.

Our problem falls exactly under this pattern.

Essentially, there are two ways to deal with goroutines here.

The first one is to communicate from a goroutine to the main thread
each time a random number has been generated and stop when we gained
enough data. In this scenario we have to listen on the channel our
goroutines write to which means main goroutine will be blocked since
reading from a channel is a blocking operation.

The second approach is about the same idea, but different
communication direction: main function defines all necessary input for
goroutines and then let them go to do their job. In this case we don't
have to block main thread every time it has to read goroutines'
result.

I went this way in my solution.

In order to implement it I set up a buffered channel `c` which is also
known as asynchronous or non-blocking channel and pushed the amount of
random numbers we need to get generated by all goroutines. Then, when
all goroutines are fired, they start to exhaust this channel one by
one until there's nothing left. Once this is done a `default`
communication switch gets triggered which simply returns from a
goroutine.