Problem 2: back pressure control :hourglass_flowing_sand:
We need to control the periods of intervals in which our goroutines
fire results.

For this Go standard library provides a set of different timers and
timer functions.

In our case we use a ticker that returns a channel and continuously
fills it with a value in a specified interval.

Delivery of values in specified intervals is guaranteed by
[non-buffered] channels' blocking nature. In other words, a goroutine
won't proceed until there's a value in the channel. And this value
appears there every second as was defined in a ticker.